Reduce memory, always set rx...
Was running into some issues with my new Trinket neopixel, where it was
not responding a lot of the time, and at times LED would come on for no
good reason, felt like memory is getting tromped.

So updated a few things:
1) Optional debug pins is now by #defines so easier to setup on
different ones (currently 9-12)

2) Updated SetRX to always update the Uart state, regardless if I did
not think it needed to be.  I was receiving data, but not processing
correct, so thought I would be safe.

3) Moved the check for diode into setRX and setTX so you don't have to
check everywhere it is called.

4) Most importantly reduced memory usage. it was 1.3K of global memory,
but now about 800 bytes, on Atemega 328 max 2K including heap/stack...
Basically removed the register write data from each regwrite item and
put it with main class, assuming last one wins, which would happen after
processing anyway.   That appeared to do the trick of making the trinket
happy.